Consistent error messages between esdocs and essql (#26399)
* Added catch promise to esdocs

* Extracted query to elasticsearch into a lib function
